requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
python-multipart
python-jose[cryptography]
passlib[bcrypt]
aiofiles
